requirements.txt
langchain
langchain-groq
fastapi
uvicorn[standard]
pydantic-settings
mysql-connector-python